+++ R/divergingx_hcl.R
@@ -37,16 +37,16 @@
 #' @param h3 numeric. Ending hue coordinate.
 #' @param c1 numeric. Chroma coordinate corresponding to \code{h1}.
 #' @param c2 numeric. Chroma coordinate corresponding to \code{h2} (if \code{NA}, set to 0).
-#' @param c3 numeric. Chroma coordinate corresponding to \code{h3}.
+#' @param c3 numeric. Chroma coordinate corresponding to \code{h3} (if \code{NA}, \code{c1} is used).
 #' @param l1 numeric. Luminance coordinate corresponding to \code{h1}.
-#' @param l2 numeric. Luminance coordinate corresponding to \code{h2}.
+#' @param l2 numeric. Luminance coordinate corresponding to \code{h2} (if \code{NA}, \code{l1} is used).
 #' @param l3 numeric. Luminance coordinate corresponding to \code{h3} (if \code{NA}, \code{l1} is used).
-#' @param p1 numeric. Power parameter for chroma coordinates in first sequential palette.
+#' @param p1 numeric. Power parameter for chroma coordinates in first sequential palette (if \code{NA}, \code{1} is used).
 #' @param p2 numeric. Power parameter for luminance coordinates in first sequential palette (if \code{NA}, \code{p1} is used).
 #' @param p3 numeric. Power parameter for chroma coordinates in second sequential palette (if \code{NA}, \code{p1} is used).
-#' @param p4 numeric. Power parameter for luminance coordinates in second sequential palette (if \code{NA}, \code{p3} is used).
+#' @param p4 numeric. Power parameter for luminance coordinates in second sequential palette (if \code{NA}, \code{p2} is used).
 #' @param cmax1 numeric. Maximum chroma coordinate in first sequential palette (not used if \code{NA}).
-#' @param cmax2 numeric. Maximum chroma coordinate in second sequential palette (if \code{NA}, \code{cmax1} is used).
+#' @param cmax2 numeric. Maximum chroma coordinate in second sequential palette (not used if \code{NA}).
 #' @param plot logical. Should the selected HCL color palettes be visualized?
 #'
 #' @return A character vector with (s)RGB codings of the colors in the palette.